requirements.txt
torch
numpy
scipy
scikit-learn
requests
pandas
matplotlib
azure-storage-blob
azure-keyvault
azure-keyvault-secrets